[error_page] Add OSE error page
:Release Notes:
Add OSE error page not requiring EnyoJS.

:Detailed Notes:
webOS OSE error page does not show because it depends on old enyojs, not
shipped in OSE. So provide a minimally modified error page not depending
on enyojs, and using ilib directly.

It adds the CSS style changes done to the chromium webos error page used
for regular error pages.

:Testing Performed:

:QA Notes:
Depends on review:210432

:Issues Addressed:
[PLAT-54756] [webOS OSE]:Network error page is not displayed after changing Date and Time

Change-Id: I5a6f6b29f431f1c6bc96044c824af8cb4edec25a