[i18n] Specify javascript file for localization
:Release Notes:
Specify javascript file for localization

:Detailed Notes:
Properties file is added for localizing strings in wam module.
Even if wam module is native application, it had used
localizatoin for web application such as $L() or getString()
So, To process this as exception, files that used it are
specified in component.properties file.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-41044] Enable Javascript file to be localized in Qt/QML or
Native application

Change-Id: I1ee51d530c7921c60c8e94e5dfbbb4d65713982c