[wam][v8context] Fixing setLaunchParams function to call updateExtensionData
:Release Notes:
The setLaunchParams function has to check if PalmSystem is ready to
receive injection data.

:Detailed Notes:
When to launch the setting apps, the updateExtensionData function is
called by setLaunchParams to inject launchParams into Palmsystem.
That's when v8context is created by the function of injection. Soon
after creating, it's disposed. This task is a burden during the launching process
for the setting app.

:Testing Performed:
Tested launching the Setting app

:QA Notes:
NA

:Issues Addressed:
[PLAT-55017] Modify the function to launch the setting app at
WAM(Integrate code)

Change-Id: I6383321deaaa5fc2a450a2c35482bc5121acfdb6