[open][emmc_opt] Better API for app preload optimizations.
:Release Notes:
Rename SetBlockWriteCache to be SetAppPreloadHint.

:Detailed Notes:
Formally WAM should not be aware of the specific optimizations done in
underlaying chromium layers, so this renames the notification passed
to WebView that app is in preload stage from something specific about
what is happening underneath (SetBlockWriteCache) to what it is, a
hint passed to WebView, so it is aware the app is in preload stage.

Then WebView, passes that information to underlaying layers, and
effectively right now it will block write to cache.

:Testing Performed:
Preloaded Youtube doesn't use disk cache.

:QA Notes:
Depends on: https://gpro.lgsvl.com/#/c/220682/
Test-cases: N/A

:Issues Addressed:
[ARCH-297] Define all launch modes and support them

Change-Id: Ie247c6078cfbd612991925a02b5290bb69482c69
Originally-Reviewed-on: http://gpro.lgsvl.com/220877