Store proxy rules in WebView profiles
Previously, the default profile was being used for all web apps. As a
result of this, the proxy rules where being shared by all web apps, and
the same tinyproxy process would be used by all of them. SMACK would
then block tinyproxy requests for all web apps but the first one, which
would be the one that configured the SMACK label in the tinyproxy
process.

Now, a mapping between app ids and WebView profiles is stored and
applied to each WebView. The proxy rules are stored on those profiles.

[SPEC-1895] error launching web app: "tinyproxy was unable to connect to the remote web server"
https://jira.automotivelinux.org/browse/SPEC-1895