Set initial size for AGL
It sets initial values for App size with AGL window size. It
couldn't get the AGL window size for now as Window manager doesn't
support that functionality for now. So, it sets the default values
for m_widthOverride and m_heightOverride as these values delivers
to WebView size and window size after getting the size value from
Env variables.

It adds 'WAM_WINDOW_SIZE' to WebAppMgr.env and passes it through
WebAppManagerConfig.

[SPEC-1890] WebApp View Size doesn't fit the homescreen.
https://jira.automotivelinux.org/browse/SPEC-1890