[agl][jsoncpp] Fix sub-objects serialization
When serializing nested objects from a JSON object
QJson behaves differently from Jsoncpp when that field
is not present in the root object. Qt returns an empty
object (string "{}") while Jsoncpp return a null value
(string "null"). This patch modifies ApplicationDescription
implementation to result in empty JSON object strings even
when that field is not present.

[SPEC-1941] WAM: jsoncpp shows warnings opening applications in log

Signed-off-by: Nick Diego Yamane <nickdiego@igalia.com>