QVariant-less WebAppWayland
This patch refactors WebAppWayland::setWindowProperty()
method, removing its dependency on QVariant. As most of
its used was using QString and in the end the value gets
convered to a std::string the solution adopted was to modify
its parameter to a std::string.

[SPEC-1911] Qt-less WAM: replace qt basic data structures

Original Author: Abhijeet Kandalkar <abhijeet@igalia.com>
Signed-off-by: Nick Diego Yamane <nickdiego@igalia.com>