QUrl-less WAM: Add Url parsing code
- The main goal of this patch is to remove WAM dependency
on QUrl implementation for Url Parsing and convenience
API. To do that, we're adding our own Url API, which in this
initial version leverages GLib URI Functions [1] for basic
URI parsing and escaping.
- Also, as the code handling URLs in WAM is mostly tied to
QFile/QDir/QFileInfo manipulation, this patch also refactor
that to use Boost Filesystem module instead. It's worth noting
it depends on a modified version os meta-agl-lge (which include
boost as dependency) to build properly. This integration is still
very primitive and will be improved soon once we have WAM build
system ported from qmake to CMake.

[SPEC-1911] Qt-less WAM: replace qt basic data structures

Signed-off-by: Nick Diego Yamane <nickdiego@igalia.com>