[agl] Register only one surface.
All the surfaces created by wam have the same pid, which
results in a wrong launcher using surfaces and activating it.

Most of the time it results in an error message and whenever
the broadcast message is received by a right launcher, the surface
is requested and activated.

A right approach will be to compare associated surface id, which
wam created for that app launcher.