[agl] React on Inactive and Active events
We need to pass to WebAppMgr host process the events, which launcher
receives. At this point, we need Inactive and Active events.

The only quick way to do it is by sending a messages through a socket
and calling the methods in WebAppManagerServiceAGL in a right
message loop through timer.

This makes the playing content to be stopped and avoid wasting
resources.