[open][wam][luna] Fix not to call pure virtual function in destructor of base class
:Release Notes:
Fix not to call pure virtual function in destructor of base class

:Detailed Notes:
In some cases, PalmServiceBase object is created for each app
and deleted when app is closed (For common webapp,
only one object is created when WebAppManager is launched)
When the object is deleted the destructor of PalmSystemBase(base class)
calls stopService to unregister service name.
And when it is failed stopService prints error message with service name
which is defined in sub class and is implemented as pure virtual
in base class. So this try to call pure virtual function causes
WebAppMgr crash sometimes.
(Actually this should be detected in compile time)
To fix this without wide change in pluggables
just changed not to print error message for error situation
If sub class wants to handle the error then should implement
stopService in its destructor to call stopService of base class
and the implemented stopService should handle error situation like
WebAppManagerServiceLuna does

:Testing Performed:
Check com.palm.webappmanager service status with
luna-send -i luna://com.webos.service.bus/signal/registerServiceStatus
'{"serviceName": "com.palm.webappmanager"}'
when webapp-mgr service is started and stopped

:QA Notes:
Test-cases: GT-27726

:Issues Addressed:
[PLAT-83557] Port wam "Luna related patches"

Change-Id: I59126ab8c748f189e27f320a09593b7120a5ea5a
Originally-Reviewed-on: http://gpro.lgsvl.com/226423
Originally-Reviewed-on: http://gpro.lge.com/250170