[open][wam][activity_manager] Delete unneccessarily created activity
:Release Notes:
wam will not create activity anymore

:Detailed Notes:
From the start of webOS, wam creates activity when web app
is launched.
However, this created activity seems not used anymore and is
redundantly registered to activity manager.
It may be a burden to activity manager so wam will not create
activity anymore.

:Testing Peformed:
Built
1. Launch any web apps
2. Perform ls-monitor -s command
3. Check there is no activity for web app at activity manager list

:QA Notes:
Test-cases: GT-27727

:Issues Addressed:
[PLAT-83557] Port wam "Luna related patches"

Change-Id: I944c13ec9adbf70a2f58bce6283546eec9a0a878
Originally-Reviewed-on: http://gpro.lgsvl.com/226454
Originally-Reviewed-on: http://gpro.lge.com/250184