[open][wam][inspector] Remove 'ENABLE_INSPECTOR' from WebAppMgr.conf
:Release Notes:
Remove 'ENABLE_INSPECTOR' from WebAppMgr.conf

:Detailed Notes:
Currently(since webOS1.0) inspector for webapp is enabled by 2 steps
1. check ENABLE_INSPECTOR which is set in WebAppMgr.conf
and do EnableDevTool if it is 1(Actually this is always 1)
This is done everytime webapp is launched(in WebPageBlink::init())
Since this is done WebAppMgr process(browser) opens 9998 port
and is listening connection request
Even after the app is closed this DevTool is still working
2. check this app is inspectable and if so then
do EnableInspectablePage for the app.
The "inspectable" of appinfo.json and
/var/lune/preferences/debug_system_apps are used for this

But because the step 1 is done in every webapp launching
even if there isn't any webapp to be inspectable
WebAppMgr is listening in 9998 port. This can be a big security hole
To fix this, remove ENABLE_INSPECTOR from WebAppMgr.conf
And just do 2 and When 2 is executed, just do EnableDevTool as well
Only when there is more than one inspectable webapp
DevTool is ready and start to listen the port 9998

:Testing Performed:
1. remove /var/luna/preferences/debug_system_apps
2. restart WebAppMgr and launch userguide
3. try to open inspector : Connection is refused
not empty 'Inspectable WebContents'
4. do netstat -nat in terminal
5. check there isn't 9998 port in LISTEN
6. do this luna api in terminal
luna-send -n 1 palm://com.palm.webappmanager/setInspectorEnable
 '{"appId":"com.webos.app.userguide"}'
7. try to open inspector
8. check there isn't any issue to use inspector for this app

1. create /var/luna/preferences/debug_system_apps
2. restart WebAppMgr and launch userguide
3. use inspector
4. there isn't any issue to use inspector

:QA Notes:
Test-cases: GT-27720, GT-27737

:Issues Addressed:
[PLAT-83525] Port chromium "[wam][inspector] EnableDevTool when set EnableInspectablePage"

Change-Id: I6117d3d1d44b834645756787808900b12748aaa4
Originally-Reviewed-on: https://gpro.lgsvl.com/226676
Originally-Reviewed-on: http://gpro.lge.com/250608