[open][pmlog] Fix use-after-free on pmlog qMessageHandler
:Release Notes:
Fix valgrind reported use-after-free on pmlog qMessageHandler.

:Detailed Notes:
This is not likely to cause any issue. But If we assign like this:
char* foo = someQstring.utf8().data()

It creates a QByteArray as a result of utf8() call. But it is not
hold and is freed immediately after this line. So the memory area
pointed by data() is already freed when sent to PmLog.

:Testing Performed:
Builds, runs.

:QA Notes:
Test-cases: N/A

:Issues Addressed:

Change-Id: I68d4aa53b25eb65f388850060777e50ffd2b636c
Originally-Reviewed-on: http://gpro.lgsvl.com/233423