[open][local_storage] Make security origin lowercase
:Release Notes:
Do not pass a security origin with uppsercase letters to Chromium.

:Detailed Notes:
When testing a web application with its appid containing uppercase
letters, that application was failing to launch, being its renderer
process killed. It was killed because of an error deserializing the mojo
IPC message for DidCommitProvisionalLoad.

Actually, the mojo message was right, but when it was read to populate
the url::Origin field, it failed because the security origin
contained uppercase letters, so constructor would reject to
construct it and would make an invalid url::Origin.

The solution provided just calls toLower() on the identifier used
to construct the security origin. It also sends a PmLog warning
in that situation.

:Testing Performed:
Launch SampleWebApp provided in PLAT-83368. Now it works.

:QA Notes:
Test-cases: GT-27739

:Issues Addressed:
[PLAT-83368] Render process crash when SampleWebApp is executed

Change-Id: I6afe55310c20c72a6176b539fc9ed7fe3841d586
Originally-Reviewed-on: http://gpro.lge.com/250618