[open][window] Support window size for system_ui
:Release Notes:
Support window size for system_ui.

:Detailed Notes:
An app with system_ui window type is supposed to have an arbitrary
window size. In case it is a web app however, it does not work as
expected even if overriding width and height from "resolution" in
appinfo.json. It is because those values are not used when creating an
application window.

:Testing Performed:
Tested with http://gpro.lge.com/253316.
1. Copy bareapp as bareapp_systemui and add items below in appinfo.json.
{
    "resolution": "400x400",
    "defaultWindowType": "system_ui"
}
2. Launch bareapp_systemui.
3. Verify that app is shown with with resolution specified.

:QA Notes:
Test-cases: QO-4344

:Issues Addressed:
[PLAT-85076] Prepare for a generic view for System UI
[PLAT-83606] Add dummy status bar sample app

Change-Id: I10ad11b11a39e6144fe3f22a039fbf5bdb8d8fbe
Originally-Reviewed-on: http://gpro.lge.com/253370