[open][util][timer] fixup timer fault in timeout callback
:Release Notes:
fixup timer fault in timeout callback

:Detailed Notes:
Timeout handler in timer callback could make timer object destroyed.
So access to timer object in timer callback should be prevented after
calling timeout handler.

eg) OneShotTimer<WebPageBlink>
      WebPageBlink::m_closeCallbackTimer

Porting it, we could not verify it actually fixes any issue, as the
crash is not happening. Though, the new code is safer so we prefer to
land the patch anyway.

:Testing Performed:
1. Add 15 seconds waiting in 'onCloseNotify' to web app.
2. Check wam crash while web app closing.
(This is not crashing right now even without the patch).

:QA Notes:
Test-cases: N/A (could not reproduce the original issue).

:Issues Addressed:
[PLAT-88718] Port "[wam][util][timer] fixup timer fault in timeout callback"

Change-Id: I5fc2e211e5d872a210c860b64a72f933973ffb51
Originally-Reviewed-on: http://gpro.lge.com/c/263155
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/263511
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>