[agl] Implement an envvar for launcher to wait for host process
When the envvar WAIT_FOR_HOST_SERVICE is set to "1", the launcher
process will wait for the shared WAM service to be enabled, instead of
attempting to run in single browser process mode.

This will be useful when webapps are auto-started on init, which could
mean that the launcher starts before the shared WAM service is ready.

For this to work, we will have to add the envvar to the corresponding
.service files for webapps, which are generated by the app framework
when a .wgt is installed.

AGL-bug: SPEC-2793

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>