Hard-code windowmanager roles for certain app ids.
Certain apps are required to have specific role names, so the window
manager assigns them to the proper region. That would be the case of
the homescreen.

This patch hard-codes the "homescreen" role to certain app ids. It's
meant to be a temporary solution to be used with the legacy window
manager, while the new one is not in place.

Bug-AGL: SPEC-2401

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>