[open][whitelist] Add media directory to the whitelist of trusted app
:Release Notes:
Add media directory to the whitelist of trusted app

:Detailed Notes:
The Home app on OSE, is a launcher app which is web app type.
The Home app need to show icon for user installed apps on the system
and the user installed apps are placed under media directory
but it didn't have permission for that directory although it was
trusted app.
After adding media directory to the whitelist for trusted app, the Home
app can show icons for installed apps as expected.

:Testing Performed:
1) Boot RPi4 device
2) Install any test app
3) Launch the Home app and verify that the icon of installed test app is
shown on the launcher.

:QA Notes:
Test-cases: QWA-1557

:Issues Addressed:
[PLAT-87058] The installed application icon is not displayed.

Change-Id: Ibe1c0433207050397c2ea7afe0037d50b37c7106
Originally-Reviewed-on: http://gpro.lge.com/259645
Reviewed-on: http://gpro.lge.com/261300
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>