[open][memory][suspend] Increase default frame deadline
:Release Notes:
Increase default frame deadline.

:Detailed Notes:
This is a provisional workaround. On coming back from suspend, after the
solution for PLAT-100262, we may see a black frame. This is because after
suspend we need to repaint the web app contents from scratch. That includes
images decoding. If that does not happen in the default deadline, then it will
emit a frame with whatever it could paint to that point. Default deadline is
4 frames (4*16.666ms = 66.666ms).

After testing a bit, we could determine a good number is 40 frames (666.666ms).
This is enough for rendering some complex pages.

Anyway, this is a workaround. On coming back from suspend we should better
ignore the deadline so it takes the time it needs to repaint all contents from
scratch.

:Testing Performed:
1. Launch app
2. Background this app
3. Open app again.

The animation to show app again should show no blank or partial web page.

:QA Notes:
Test-cases: GT-28309
Depends on http://gpro.lge.com/270394

:Issues Addressed:
[PLAT-100262] [chromium79] Chromium79 uses more GPU memory compared to chromium72

Change-Id: I23c334d50712bce3996268dee2b6b24777a6a0b5
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/270708
Reviewed-by: Build_Verification
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>