[op][error_page] Reload the current page for ERR_NETWORK_CHANGED
:Release Notes:
Reload the current page for ERR_NETWORK_CHANGED rather than showing error page.

:Detailed Notes:
For OSE and Auto, ERR_NETWORK_CHANGED is returned from chromium when
the network is reconnected.
Since the network is changed to normal, it is abnormal to show it as a
error page.
So rather than showing the error page, reload the current page.

:Testing Performed:
1) rm -rf /var/lib/wam/*
2) Disconnect the network cable
3) Launch YouTube test app
4) When -106 error page is shown, connect the network cable
5) Verify that YouTube page is successfully loaded

:QA Notes:
Test-cases: QWA-1055

:Issues Addressed:
[PLAT-103584] [AUTO] Network recovery is failed.

Change-Id: Ia5d6559ecf371f51fdfecad918202490fcdd5775
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/275056
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>