[op][tellurium] Reload Tellurium after URL redirection
:Release Notes:
Reload Tellurium after URL redirection

:Detailed Notes:
There is a problem that Tellurium is not loaded properly when URL is
redirected after the document load is finished.
This is because the m_userScriptExecuted is not initalized and the
reason is that BlinkWebView::LoadStarted() is invoked only once even
though the document is newly loaded with the redirection URL.

To fix this issue, change the place for initializing
m_userScriptExecuted to DidStartNavigation() from LoadStarted().

:Testing Performed:
1) Install test.redirection_1.0.1_all.ipk
2) Launch test.redirection app with below command.
"test.redirection", "params":{"target":"http://10.177.245.57/~gabwon.park/web/html5_input_test.html"}}'
3) Verify that test.rediction is visible on tellurium service with below
command.
'{"command": "getRunningAppsAndStages"}'

:QA Notes:
Test-cases: QS-8715

:Issues Addressed:
[PLAT-111247] [Signage] Test Redirection app in not showing up in
com.palm.service.tellurium/handleRequest {getRunningAppsAndStages}

Change-Id: I8c3620bb6d54c25b38519ef7d54b02f36b22dbff
Reviewed-on: http://gpro.lge.com/c/webos-pro/wam/+/279547
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/279546