[op][media] Drop video texture API
:Release Notes:
Drop unused video texture API

:Detailed Notes:
The API to notify Chromium that an application rotation was 0º, so it
would force video texture media playback was not used at all. This patch
just removes the usage of that API (that should also be removed in
Chromium side.

:Testing Performed:
Build completes.

:QA Notes:
Test-cases: not required (only build).
Depends-on: http://gpro.lge.com/c/webosose/chromium79/+/289491

:Issues Addressed:
[PLAT-119527] Drop unused code for application life cycle that could make splash screen refactor harder

Change-Id: I955706fc62595beca95ae55f0334ac9bddcfdfeb
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/289494
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Static Analysis <dtvqe@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Jose Dapena <jose.dapena@lge.com>