[op][qtless][life_cycle] Unregister from running list apps on closing.
:Release Notes:
Prevent crashes when applications are closed without a WAM request, accessing running list.

:Detailed Notes:
WAM crashes accessing the list of running applications, if an application
has been closed, but not because of WAM (i.e. after renderer process crash).
This is because in those cases the WebAppBase instance is not removed from
the list of running  apps.

To avoid that situation, we explicitely call AppDeleted in WebAppManager
also in those situations. It may be redundant if it was already called,
but it will still be valid.

:Testing Performed:
1. Kill home application
2. Call luna://com.webos.service.applicationmanager/running

:QA Notes:
Test-cases: QWT-472, QWT-4148

:Issues Addressed:
[WRN-11100] [RDX_CRASH][webos] /usr/bin/WebAppMgr in _ZNK12WebPageBlink16RenderProcessPidEv (libWebAppMgr.so.1.0 + 0x2ff20)
[WRN-12118] [RDX_CRASH][webos] /usr/bin/WebAppMgr in _ZNK10WebAppBase10InstanceIdB5cxx11Ev (libWebAppMgrCore.so.1.0 + 0x25094)
[WRN-12163] [RDX_CRASH][webos] /usr/bin/WebAppMgr in _ZNK12WebPageBlink16RenderProcessPidEv (libWebAppMgr.so.1.0 + 0x2ff24)

Change-Id: I52c95d85e5f4229ef4c58c7792054e45443b4fe0
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/321837
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>