[op][life_cycle] Improve checks for killApp reasons.
:Release Notes:
Avoid looping forever on a request to kill a web application from
memory manager.

:Detailed Notes:
In webOS OSE, SAM requests to kill an application because of memory
manager have as reason "com.webos.service.memorymanager" instead of
"memoryReclaim". But, also, when SAM asks to close a non keep-alive
application, and it takes more than a second, attempts will happen
without a reason field value.

In both cases, this change just forces closing the web application
immediately.

:Testing Performed:
Force an application to close because of memory usage using memstay
and launching several apps. The attempts to close should happen every
second, and application should be closed and finalized soon, even
if it is a keep alive application.

:QA Notes:
Depends on: http://gpro.lge.com/c/webosose/sam/+/321872

:Issues Addressed:
[WRN-12866] [RDX_CRASH][webos] /usr/bin/WebAppMgr in n/a (libcbe.so + 0x313a304)

Change-Id: Ib17c8b0180b9c6978b6d3dca959cf3d21c49965a
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/321873
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>