[op][n_upsable][media] Pass "enableBackgroundRun" appinfo property
:Release Notes:
Expose "enableBackgroundRun" appinfo property value to Chromium

:Detailed Notes:
With the latest patch set: http://gpro.lge.com/c/333721
chromium uses upstream provided suspend/resume framework using VDA.
Which causes media playback to suspend when application goes to
background.

But for some cases media playback should not be suspended even
when app goes to background. With this patch set we pass the
value of appinfo property "enableBackgroundRun" to Chromium.
And chromium checks for the value of appinfo property
and does not suspend media playback for the web app if it
is in background and if appinfo value is enabled.

:Testing Performed:
Below test cases are verified.

:QA Notes:
Test cases: QWTC-435 , QWTC-436

Depends On: http://gpro.lge.com/c/webosose/chromium91/+/334559

:Issues Addressed:
[WRO-10156] [OSE -RPI4]:Audio not played via BT Headset after launching Test redirection app

Change-Id: I017a7bc61ff8936df95b56840651bd7c62ed2ba5
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/334550
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Praveen Kumar P <praveen.p@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Praveen Kumar P <praveen.p@lge.com>