[op][webpush] Update the permission status for notification
:Release Notes:
Update the permission status for notification

:Detailed Notes:
In this patch, A web application manager(WAM) can subscribe to the
service and deliver the app id and its permission status to settings in
web runtime whenever wam is initialized e or the ststus is changed.
We assume that there is a service to provide app id and its permission
status in webOS. So, there needs additional implementation for
interacting with the service provided on each platform.

:Testing Performed:
1. rm /var/lib/wam/Default/wam_prefs.json
2. restart webapp-mgr
3. wait for 10 sec
4. ls /var/lib/wam/Default/wam_prefs.json
5. cat /var/lib/wam/Default/wam_prefs.json

expected result:
"notifications":{
    "file://test.web-push-webos,*":{"..., "setting":1}
}

:QA Notes:
Test-cases: QWTC-8657, QWTC-13686
Depends on: http://gpro.lge.com/c/webosose/chromium108/+/379317

:Issues Addressed:
[WRP-8225] Patch migration from webos-pro/wam (web push)
[WRO-12745] permission for platform
[WRP-1097] Delete web push permission info when app is deinstalled
[WRQ-1411] Review and migrate web push patches (wam/webruntime)

Change-Id: I143c326382f6594041fb0d1f2f8fbc15a7f92ea0
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/345120
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/380351
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Blink-Build: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>