[op][tests] Fix failing unit tests after upgrading to Chromium94
:Release Notes:
Fix failing unit tests after upgrading to Chromium94

:Detailed Notes:
Some unit tests are failing after upgrading to Chromium94.
Most of them was because of the change of the url of file scheme with
below commit:
http://gpro.lge.com/c/webosose/wam/+/333697

Apply proper fix to related unit tests.

:Testing Performed:
Run unit test and verify that all TCs are passed.

:QA Notes:
Test-cases: QWTC-9981

:Issues Addressed:
[WRO-14002] [OSE-RPi4]:  1 Failure is displayed for WAM test binary

Change-Id: I6395eb5e4fbe81bd525824ff5269635a020c6d6d
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/338011
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>