fixup! [op][qtless] Move to chromium coding style.
:Release Notes:
Use std::string() instead of nullptr for const std::string&.

:Detailed Notes:
The second parameter of SetGroupClientEnvironment() is not used
when the first parameter is not kKeyMask.
However, we should provide a dummy value that does not make a crash.

:Testing Performed:
Evaluate PalmSystem.focusOwner() in a web app.
No crash should happen.

:QA Notes:
Test-cases: QWTC-3996

:Issues Addressed:
[WRP-10810] com.webos.exampleapp.enyoapp.epg app closes automatically.

Change-Id: I6b0f16fe4753902ba8676c82983e5e57382604d7
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/348230
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>