[op][error_page] Remove a useless parameter related to WebPageBlink::loadFailed
:Release Notes:
Remove a useless parameter which is always the empty string

:Detailed Notes:
Since Chromium68, the third argument of WebViewDelegate::LoadFailed has
been the empty string.
In the past versions, WebView::DidFinishNavigation called
WebView::DidFailLoad with base::ASCIIToUTF16(""), and
WebView::DidFailLoad forwarded base::UTF16ToUTF8(error_description) to
WebViewDelegate::LoadFailed.
Recently, WebView::DispatchNetError has called
WebViewDelegate::LoadFailed with std::string().
The value of the error description is always the empty string, and this
value seems useless in WAM. Specifically, the value of the error
description is just used to compare it with "SSL_ERROR" in
WebPageBlink::loadFailed.
However, the actual value of the error description is always "".
For these reasons, the parameter, const std::string& errDesc,
would be removed.

:Testing Performed:
Build OK

:QA Notes:
Test-cases: N/A
Depends-on: http://gpro.lge.com/c/webosose/chromium108/+/390499

:Issues Addressed:
[[WRQ-8302] [error_page] Update some commits related to an error page
[WRO-13028] [setLoadErrorPolicy]Unable to move to network error page when Load URL
[WRP-7711] Restore "fixup! [op][error_page] Remove a useless parameter
           related to WebPageBlink::loadFailed"

Change-Id: Ifb06d1b4813445dbe7583fdf87715c9f5727a8b5
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/345054
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/390497
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Jeongran Jang <jeongran.jang@lge.com>