[code health] Simplify ownership of ApplicationDescription
:Release Notes:
Simplify ownership of ApplicationDescription.

:Detailed Notes:
ApplicationDescription is shared by WebAppBasePrivate and WebPageBase.
Make WebAppBasePrivate own it, and make WebPageBase have a reference of
it.

Because CreateWebApp() and CreateWebPage() of WebAppFactoryInterface do
not transfer the ownership, make them to use the reference.

Remove WebAppBase::CleanResources, WebPageBase::WebPageBase() and
WebPageBase::SetApplicationDescription().

:Testing Performed:
Build successfully.

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRR-537] Check about std::shared_ptr<ApplicationDescription>

Change-Id: Ibff55dc9e3a29a95e65c8c992a3faaec5ce53576
Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/431881
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>